+++ statistics-sdk/src/main/java/com/dwarfeng/statistics/sdk/bean/entity/JSFixedFastJsonHistoryTask.java
@@ -17,7 +17,7 @@ import java.util.Objects;
  */
 public class JSFixedFastJsonHistoryTask implements Bean {
 
-    private static final long serialVersionUID = 3092831961673541171L;
+    private static final long serialVersionUID = -6620117446335946098L;
 
     public static JSFixedFastJsonHistoryTask of(HistoryTask historyTask) {
         if (Objects.isNull(historyTask)) {
@@ -34,7 +34,7 @@ public class JSFixedFastJsonHistoryTask implements Bean {
                     historyTask.getExpiredDate(),
                     historyTask.getDiedDate(),
                     historyTask.getExecutionNodeId(),
-                    historyTask.getFrontMessage(),
+                    historyTask.getMessage(),
                     historyTask.getRemark()
             );
         }
@@ -70,8 +70,8 @@ public class JSFixedFastJsonHistoryTask implements Bean {
     @JSONField(name = "execution_node_id", ordinal = 10)
     private int executionNodeId;
 
-    @JSONField(name = "front_message", ordinal = 11)
-    private String frontMessage;
+    @JSONField(name = "message", ordinal = 11)
+    private String message;
 
     @JSONField(name = "remark", ordinal = 12)
     private String remark;
@@ -82,7 +82,7 @@ public class JSFixedFastJsonHistoryTask implements Bean {
     public JSFixedFastJsonHistoryTask(
             JSFixedFastJsonLongIdKey key, JSFixedFastJsonLongIdKey statisticsSettingKey, int status, Date createDate,
             Date startDate, Date endDate, Long duration, Date expiredDate, Date diedDate, int executionNodeId,
-            String frontMessage, String remark
+            String message, String remark
     ) {
         this.key = key;
         this.statisticsSettingKey = statisticsSettingKey;
@@ -94,7 +94,7 @@ public class JSFixedFastJsonHistoryTask implements Bean {
         this.expiredDate = expiredDate;
         this.diedDate = diedDate;
         this.executionNodeId = executionNodeId;
-        this.frontMessage = frontMessage;
+        this.message = message;
         this.remark = remark;
     }
 
@@ -178,12 +178,12 @@ public class JSFixedFastJsonHistoryTask implements Bean {
         this.executionNodeId = executionNodeId;
     }
 
-    public String getFrontMessage() {
-        return frontMessage;
+    public String getMessage() {
+        return message;
     }
 
-    public void setFrontMessage(String frontMessage) {
-        this.frontMessage = frontMessage;
+    public void setMessage(String message) {
+        this.message = message;
     }
 
     public String getRemark() {
@@ -207,7 +207,7 @@ public class JSFixedFastJsonHistoryTask implements Bean {
                 ", expiredDate=" + expiredDate +
                 ", diedDate=" + diedDate +
                 ", executionNodeId=" + executionNodeId +
-                ", frontMessage='" + frontMessage + '\'' +
+                ", message='" + message + '\'' +
                 ", remark='" + remark + '\'' +
                 '}';
     }