+++ statistics-stack/src/main/java/com/dwarfeng/statistics/stack/bean/entity/Task.java
@@ -13,7 +13,7 @@ import java.util.Date;
  */
 public class Task implements Entity<LongIdKey> {
 
-    private static final long serialVersionUID = -6608818783384186394L;
+    private static final long serialVersionUID = 8716595217841154300L;
 
     private LongIdKey key;
     private LongIdKey statisticsSettingKey;
@@ -44,7 +44,7 @@ public class Task implements Entity<LongIdKey> {
     /**
      * 任务执行是在最前方显示的信息。
      */
-    private String frontMessage;
+    private String message;
 
     private String remark;
 
@@ -53,7 +53,7 @@ public class Task implements Entity<LongIdKey> {
 
     public Task(
             LongIdKey key, LongIdKey statisticsSettingKey, int status, Date createDate, Date startDate,
-            Date shouldExpireDate, Date shouldDieDate, int executionNodeId, String frontMessage, String remark
+            Date shouldExpireDate, Date shouldDieDate, int executionNodeId, String message, String remark
     ) {
         this.key = key;
         this.statisticsSettingKey = statisticsSettingKey;
@@ -63,7 +63,7 @@ public class Task implements Entity<LongIdKey> {
         this.shouldExpireDate = shouldExpireDate;
         this.shouldDieDate = shouldDieDate;
         this.executionNodeId = executionNodeId;
-        this.frontMessage = frontMessage;
+        this.message = message;
         this.remark = remark;
     }
 
@@ -133,12 +133,12 @@ public class Task implements Entity<LongIdKey> {
         this.executionNodeId = executionNodeId;
     }
 
-    public String getFrontMessage() {
-        return frontMessage;
+    public String getMessage() {
+        return message;
     }
 
-    public void setFrontMessage(String frontMessage) {
-        this.frontMessage = frontMessage;
+    public void setMessage(String message) {
+        this.message = message;
     }
 
     public String getRemark() {
@@ -160,7 +160,7 @@ public class Task implements Entity<LongIdKey> {
                 ", shouldExpireDate=" + shouldExpireDate +
                 ", shouldDieDate=" + shouldDieDate +
                 ", executionNodeId=" + executionNodeId +
-                ", frontMessage='" + frontMessage + '\'' +
+                ", message='" + message + '\'' +
                 ", remark='" + remark + '\'' +
                 '}';
     }